Move Sync examples out of the experimental namespace
