+++ .github/workflows/c++.yml
@@ -21,8 +21,8 @@ jobs:
               - 'examples/cpp/beta/asymmetric/**'
             local:
               - 'examples/cpp/local/**'
-            beta-sync:
-              - 'examples/cpp/beta/sync/**'
+            sync:
+              - 'examples/cpp/sync/**'
 
       - name: Realm C++ Beta Asymmetric Tests
         if: steps.filter.outputs.beta-asymmetric == 'true'
@@ -39,12 +39,12 @@ jobs:
           mkdir build && cd build
           cmake ..
           cmake --build .
-          ./examples-beta-local
-      - name: Realm C++ Beta Flexible Sync Tests
-        if: steps.filter.outputs.beta-sync == 'true'
+          ./examples-local
+      - name: Realm C++ Sync Tests
+        if: steps.filter.outputs.sync == 'true'
         run: |
-          cd examples/cpp/beta/sync
+          cd examples/cpp/sync
           mkdir build && cd build
           cmake ..
           make
-          ./examples-beta-sync
+          ./examples-sync
