--- examples/cpp/beta/sync/call-function.cpp
@@ -1,37 +1,0 @@
-#include <catch2/catch_test_macros.hpp>
-#include <string>
-#include <future>
-#include <cpprealm/sdk.hpp>
-
-static const std::string APP_ID = "cpp-tester-uliix";
-
-TEST_CASE("call a function", "[realm][sync]")
-{
-    // :snippet-start: beta-call-a-function
-    // Connect to an App Services App and authenticate a user
-    // :snippet-start: connect-app-services
-    auto appConfig = realm::App::configuration();
-    appConfig.app_id = APP_ID;
-    auto app = realm::App(appConfig);
-    // :snippet-end:
-    auto user = app.login(realm::App::credentials::anonymous()).get();
-    auto sync_config = user.flexible_sync_configuration();
-
-    // If a function takes arguments, pass them as BSON
-    auto arg1 = realm::bson::Bson("john.smith");
-    auto arg2 = realm::bson::Bson("@companyemail.com");
-
-    // Call an App Services function as the logged-in user
-    auto result = user.call_function("concatenate", { arg1, arg2 }).get();
-
-    // Verify that the result has a value
-    CHECK(result);
-    auto bsonResult = result.value();
-
-    // Translate the BSON result back to a string
-    auto resultString = std::string(bsonResult);
-    // Prints "Calling the concatenate function returned john.smith@companyemail.com."
-    std::cout << "Calling the concatenate function returned " << resultString << ".\n";
-    // :snippet-end:
-    REQUIRE(resultString == "john.smith@companyemail.com");
-}
